WIP almost got everything hooked up. Missing TextureViews and descriptor bindings for render

+++ src/Vex/Bindings.h
@@ -20,7 +20,7 @@ struct ConstantBinding
     }
 
     void* data;
-    size_t size;
+    u32 size;
 };
 
 // clang-format off
